requirements.txt
numpy
pandas
torch
torch-geometric
scikit-learn
PyYAML